refactor requests to the proxy
