+++ applications/randomx_miner/src/run_miner.rs
@@ -22,23 +22,19 @@
 
 use std::{sync::Arc, time::Duration};
 
-use log::{debug, error, info};
-use minotari_app_utilities::parse_miner_input::wallet_payment_address;
+use log::info;
 use reqwest::Client;
-use serde::{Deserialize, Serialize};
-use serde_json::json;
 use tari_common::{load_configuration, DefaultConfigLoader};
 use tokio::{sync::Mutex, time::sleep};
 
 use crate::{
     cli::Cli,
     config::RandomXMinerConfig,
-    error::{ConfigError, Error, RequestError},
-    json_rpc::{GetBlockCountResponse, GetBlockTemplateResponse, Request},
+    error::{ConfigError, Error},
+    json_rpc::{get_block_count, get_block_template},
 };
 
 pub const LOG_TARGET: &str = "minotari::randomx_miner::main";
-pub const LOG_TARGET_FILE: &str = "minotari::logging::randomx_miner::main";
 
 pub async fn start_miner(cli: Cli) -> Result<(), Error> {
     let config_path = cli.common.config_path();
@@ -61,66 +57,9 @@ pub async fn start_miner(cli: Cli) -> Result<(), Error> {
     loop {
         info!(target: LOG_TARGET, "Starting new mining cycle");
 
-        get_tip_info(&client, &node_address, tip.clone()).await?;
-        get_block_template(&client, &node_address, &monero_wallet_address).await?;
+        get_block_count(&client, &node_address, tip.clone()).await?;
+        let block_template = get_block_template(&client, &node_address, &monero_wallet_address).await?;
+
         sleep(Duration::from_secs(15)).await
     }
-}
-
-async fn get_tip_info(client: &Client, node_address: &String, tip: Arc<Mutex<u64>>) -> Result<(), Error> {
-    let response = client
-        .post(format!("{}/json_rpc", &node_address.to_string())) // Replace with your node's address
-        .json(&Request::new("get_block_count", serde_json::Value::Null))
-        .send().await.map_err(|e| {
-        error!(target: LOG_TARGET, "Reqwest error: {:?}", e);
-        Error::from(RequestError::GetBlockCount(e.to_string()))
-    })?
-        .json::<GetBlockCountResponse>().await?;
-    debug!(target: LOG_TARGET, "`get_block_count` Response: {:?}", response);
-
-    if response.result.status == "OK" {
-        debug!(target: LOG_TARGET, "`get_block_count` Blockchain tip (block height): {}", response.result.count);
-        *tip.lock().await = response.result.count;
-    } else {
-        debug!(target: LOG_TARGET, "Failed to get the block count. Status: {}", response.result.status);
-        return Err(RequestError::GetBlockCount(format!(
-            "Failed to get the block count. Status: {}",
-            response.result.status
-        ))
-        .into());
-    }
-
-    Ok(())
-}
-
-async fn get_block_template(
-    client: &Client,
-    node_address: &String,
-    monero_wallet_address: &String,
-) -> Result<(), Error> {
-    let response = client
-            .post(format!("{}/json_rpc", &node_address.to_string())) // Replace with your node's address
-            .json(&Request::new("get_block_template", json!({
-                "wallet_address": monero_wallet_address,
-                "reserve_size": 60,
-            })))
-            .send().await.map_err(|e| {
-            error!(target: LOG_TARGET, "Reqwest error: {:?}", e);
-            Error::from(RequestError::GetBlockTemplate(e.to_string()))
-        })?
-            .json::<GetBlockTemplateResponse>().await?;
-    debug!(target: LOG_TARGET, "`get_block_template` Response: {:?}", response);
-
-    if response.result.status == "OK" {
-        debug!(target: LOG_TARGET, "`get_block_template` Block template: {:?}", response.result);
-    } else {
-        debug!(target: LOG_TARGET, "Failed to get the block template. Status: {}", response.result.status);
-        return Err(RequestError::GetBlockCount(format!(
-            "Failed to get the block template. Status: {}",
-            response.result.status
-        ))
-        .into());
-    }
-
-    Ok(())
 }
